Update PyPI classifiers to supported Py versions
Recent commits changed CI to only run Python 3.6 through 3.9:
- 3beb813b9b5a553b198b7b33b469b7bb7f2b231b
- 180008235aa2777e0157ea2a784c996d1e73de8d

Additionally, the 0.2.6 release states that only versions 3.6 and up are
supported.

Python 2.7 and 3.5 are both EOL (Python 3.6 will lose support in
December of this year), so it's helpful to not support them anymore.

Adding these classifiers can help those who are inspecting the package
on PyPI and trying to figure out which versions of Python they can use.